test(SV-BREAKPOINT-001): fuzz tests for split-read breakpoint detection
Fuzz the FindBreakpoints/RefineBreakpoint contract over the BE boundaries
(identical junctions, single breakpoint below the support gate, no split
reads, 0/-1/MaxInt coordinates) plus the §7.1 worked example and a randomized
sweep with an independent spec recomputation.

The randomized sweep surfaced an OverflowException in FindBreakpoints: the
cluster-gap test computed Math.Abs(curr - prev) in Int32, which overflows
(Math.Abs of int.MinValue) on extreme/opposite-sign junction coordinates.
Per the §4.1 positional-gap definition the comparison is now done in 64-bit
width so extreme reference coordinates classify correctly without throwing.

+++ src/Seqeron/Algorithms/Seqeron.Genomics.Annotation/StructuralVariantAnalyzer.cs
@@ -607,9 +607,12 @@ public static class StructuralVariantAnalyzer
             var curr = reads[i];
 
             // Same breakpoint iff same chromosome and junction within the tolerance window.
+            // The gap is computed in 64-bit width: junctions are reference coordinates that
+            // may span the full Int32 range, and an Int32 subtraction (or Math.Abs of
+            // int.MinValue) would overflow and throw on extreme/opposite-sign coordinates.
             bool sameCluster =
                 prev.Chromosome == curr.Chromosome &&
-                Math.Abs(curr.SupplementaryPosition - prev.SupplementaryPosition) <= clusterTolerance;
+                Math.Abs((long)curr.SupplementaryPosition - prev.SupplementaryPosition) <= clusterTolerance;
 
             if (sameCluster)
             {
